+++ src/Hangfire.Mongo/MongoFetchedJob.cs
@@ -28,11 +28,11 @@ namespace Hangfire.Mongo
         /// <param name="id">Identifier</param>
         /// <param name="jobId">Job ID</param>
         /// <param name="queue">Queue name</param>
-        public MongoFetchedJob(HangfireDbContext database, ObjectId id, string jobId, string queue)
+        public MongoFetchedJob(HangfireDbContext database, ObjectId id, ObjectId jobId, string queue)
         {
             _database = database ?? throw new ArgumentNullException(nameof(database));
             _id = id;
-            JobId = jobId ?? throw new ArgumentNullException(nameof(jobId));
+            JobId = jobId.ToString();
             Queue = queue ?? throw new ArgumentNullException(nameof(queue));
         }