+++ src/Hangfire.Mongo/MongoMonitoringApi.cs
@@ -85,19 +85,21 @@ namespace Hangfire.Mongo
         {
             return UseConnection(database =>
             {
-                JobDto job = database.Job.Find(Builders<JobDto>.Filter.Eq(_ => _.Id, jobId))
+                JobDto job = database.Job.Find(Builders<JobDto>.Filter.Eq(_ => _.Id, ObjectId.Parse(jobId)))
                     .FirstOrDefault();
 
                 if (job == null)
                     return null;
 
                 var history = job.StateHistory.Select(x => new StateHistoryDto
-                {
-                    StateName = x.Name,
-                    CreatedAt = x.CreatedAt,
-                    Reason = x.Reason,
-                    Data = x.Data
-                }).ToList();
+                    {
+                        StateName = x.Name,
+                        CreatedAt = x.CreatedAt,
+                        Reason = x.Reason,
+                        Data = x.Data
+                    })
+                    .Reverse()
+                    .ToList();
 
                 return new JobDetailsDto
                 {
@@ -128,12 +130,12 @@ namespace Hangfire.Mongo
                     .Group(dto => new { dto.StateName }, dtos => new { StateName = dtos.First().StateName, Count = dtos.Count() })
                     .ToList().ToDictionary(kv => kv.StateName, kv => kv.Count);
 
-                Func<string, int> getCountIfExists = name => countByStates.ContainsKey(name) ? countByStates[name] : 0;
+                int GetCountIfExists(string name) => countByStates.ContainsKey(name) ? countByStates[name] : 0;
 
-                stats.Enqueued = getCountIfExists(EnqueuedState.StateName);
-                stats.Failed = getCountIfExists(FailedState.StateName);
-                stats.Processing = getCountIfExists(ProcessingState.StateName);
-                stats.Scheduled = getCountIfExists(ScheduledState.StateName);
+                stats.Enqueued = GetCountIfExists(EnqueuedState.StateName);
+                stats.Failed = GetCountIfExists(FailedState.StateName);
+                stats.Processing = GetCountIfExists(ProcessingState.StateName);
+                stats.Scheduled = GetCountIfExists(ScheduledState.StateName);
 
                 stats.Servers = database.Server.Count(new BsonDocument());
 
@@ -318,8 +320,9 @@ namespace Hangfire.Mongo
 
         private JobList<EnqueuedJobDto> EnqueuedJobs(HangfireDbContext database, IEnumerable<string> jobIds)
         {
+            var jobObjectIds = jobIds.Select(ObjectId.Parse);
             var jobs = database.Job
-                .Find(Builders<JobDto>.Filter.In(_ => _.Id, jobIds))
+                .Find(Builders<JobDto>.Filter.In(_ => _.Id, jobObjectIds))
                 .ToList();
 
             var filterBuilder = Builders<JobQueueDto>.Filter;
@@ -371,7 +374,7 @@ namespace Hangfire.Mongo
             {
                 var stateData = job.StateData;
                 var dto = selector(job, DeserializeJob(job.InvocationData, job.Arguments), stateData);
-                result.Add(new KeyValuePair<string, TDto>(job.Id, dto));
+                result.Add(new KeyValuePair<string, TDto>(job.Id.ToString(), dto));
             }
 
             return new JobList<TDto>(result);
@@ -402,8 +405,9 @@ namespace Hangfire.Mongo
 
         private JobList<FetchedJobDto> FetchedJobs(HangfireDbContext database, IEnumerable<string> jobIds)
         {
+            var jobObjectIds = jobIds.Select(ObjectId.Parse);
             var jobs = database.Job
-                .Find(Builders<JobDto>.Filter.In(_ => _.Id, jobIds))
+                .Find(Builders<JobDto>.Filter.In(_ => _.Id, jobObjectIds))
                 .ToList();
 
             var jobIdToJobQueueMap = database.JobQueue
@@ -438,7 +442,7 @@ namespace Hangfire.Mongo
             foreach (var job in joinedJobs)
             {
                 result.Add(new KeyValuePair<string, FetchedJobDto>(
-                    job.Id,
+                    job.Id.ToString(),
                     new FetchedJobDto
                     {
                         Job = DeserializeJob(job.InvocationData, job.Arguments),
@@ -450,7 +454,7 @@ namespace Hangfire.Mongo
             return new JobList<FetchedJobDto>(result);
         }
 
-        private JobList<TDto> GetJobs<TDto>(HangfireDbContext database, int from, int count, string stateName, Func<JobDetailedDto, Job, Dictionary<string, string>, TDto> selector)
+        private static JobList<TDto> GetJobs<TDto>(HangfireDbContext database, int from, int count, string stateName, Func<JobDetailedDto, Job, Dictionary<string, string>, TDto> selector)
         {
             // only retrieve job ids
             var filter = Builders<JobDto>
@@ -459,6 +463,7 @@ namespace Hangfire.Mongo
 
             var jobs = database.Job
                 .Find(filter)
+                .SortByDescending(_ => _.Id)
                 .Skip(from)
                 .Limit(count)
                 .ToList();